+++ TaskMaster/include/MainPadge/MainPadgeHandlerInterface.hpp
@@ -7,8 +7,8 @@ class MainPadgeHandlerInterface
 {
 public:
     virtual ~MainPadgeHandlerInterface() {}
-    virtual std::vector<Project> GetUserProjects(int userId) = 0;
-    virtual Project GetProjectById(int projectId) = 0;
-    virtual bool CreateProject(int userId, const std::string &projectName) = 0;
+    virtual std::vector<Project> GetUserProjects(size_t userId) = 0;
+    virtual Project GetProjectById(size_t projectId) = 0;
+    virtual bool CreateProject(size_t userId, const std::string &projectName) = 0;
 };
