+++ TaskMaster/include/Project/ProjectController.hpp
@@ -19,10 +19,10 @@ class ProjectController : public ProjectHandlerInterface
 public:
     ProjectController(IProjectRepoUptr projectRepo, IBoardRepoUptr boardRepo, IUserRepoUptr userRepo);
     ~ProjectController();
-    void CreateBoard(int projectId, int creatorId, const std::string &boardName, const std::string &text) override;
-    void AddUser(int projectId, const std::string &userName) override;
-    std::vector<User> GetAllUsers(int projectId) override;
-    std::vector<Board> GetAllBoards(int projectId) override;
+    void CreateBoard(size_t projectId, size_t creatorId, const std::string &boardName, const std::string &text) override;
+    void AddUser(size_t projectId, const std::string &userName) override;
+    std::vector<User> GetAllUsers(size_t projectId) override;
+    std::vector<Board> GetAllBoards(size_t projectId) override;
 
 private:
     IProjectRepoUptr projectRepo_;
