+++ TaskMaster/include/Project/ProjectHandlerInterface.hpp
@@ -9,8 +9,8 @@ class ProjectHandlerInterface
 {
 public:
     virtual ~ProjectHandlerInterface() {}
-    virtual void CreateBoard(int projectId, int creatorId, const std::string &boardName, const std::string &text) = 0;
-    virtual void AddUser(int projectId, const std::string &userName) = 0;
-    virtual std::vector<User> GetAllUsers(int projectId) = 0;
-    virtual std::vector<Board> GetAllBoards(int projectId) = 0;
+    virtual void CreateBoard(size_t projectId, size_t creatorId, const std::string &boardName, const std::string &text) = 0;
+    virtual void AddUser(size_t projectId, const std::string &userName) = 0;
+    virtual std::vector<User> GetAllUsers(size_t projectId) = 0;
+    virtual std::vector<Board> GetAllBoards(size_t projectId) = 0;
 };
