+++ TaskMaster/include/Project/ProjectRepoInterface.hpp
@@ -6,10 +6,10 @@ class ProjectRepoInterface
 {
 public:
     virtual ~ProjectRepoInterface() {}
-    virtual std::vector<Project> GetUserProjects(int userId) = 0;
-    virtual Project GetProjectById(int projectId) = 0;
-    virtual bool CreateProject(int userId, const std::string &projectName) = 0;
-    virtual bool RemoveProjectById(int projectId) = 0;
-    virtual bool AddUserToProject(int projectId, const std::string &userName) = 0;
+    virtual std::vector<Project> GetUserProjects(size_t userId) = 0;
+    virtual Project GetProjectById(size_t projectId) = 0;
+    virtual bool CreateProject(size_t userId, const std::string &projectName) = 0;
+    virtual bool RemoveProjectById(size_t projectId) = 0;
+    virtual bool AddUserToProject(size_t projectId, const std::string &userName) = 0;
     virtual bool UpdateProject(const Project& project) = 0;
 };
