+++ TaskMaster/src/MainPadge/MainPadgeController.cpp
@@ -2,18 +2,17 @@
 
 MainPadgeController::MainPadgeController(IProjectRepoUptr projectRepo) : projectRepo_(std::move(projectRepo)) {}
 
-std::vector<Project> MainPadgeController::GetUserProjects(int userId)
+std::vector<Project> MainPadgeController::GetUserProjects(size_t userId)
 {
-    // проверка на валидность тут?
     return projectRepo_->GetUserProjects(userId);
 }
 
-Project MainPadgeController::GetProjectById(int projectId)
+Project MainPadgeController::GetProjectById(size_t projectId)
 {
     return projectRepo_->GetProjectById(projectId);
 }
 
-bool MainPadgeController::CreateProject(int userId, const std::string &projectName)
+bool MainPadgeController::CreateProject(size_t userId, const std::string &projectName)
 {
     return projectRepo_->CreateProject(userId, projectName);
 }
